Job Scheduling Algorithms_MSO_Sem2
Comparative analysis of Integer Programming (IP), Genetic Algorithm (GA), and Ant Colony Optimization (ACO) to solve the Job scheduling problem (JSP). The results showcased that IP gave the exact optimal solution and GA also provided the same sequence along with same processing time. ACO showed same processing time as that of IP and GA but with a difference sequence of jobs over the machines. ACO proved that its flexible to adapt itself to the changing environments and can handle complex problems. This analysis explains the balance between all the three algorithms highlighting the advantages and disadvantages among which scalability and flexibility is crucial.